revert: reset `COMPILER_RT_USE_BUILTINS_LIBRARY` to "ON", and disable building memprof
we reset option `COMPILER_RT_USE_BUILTINS_LIBRARY` to "ON", which will use compiler-rt builtins,
instead of libgcc.
we also discussed the value settings of this option in PR #1.
see also #1 and commit e080b38.

In llvm compile-rt, a new component `memprof` was added, we need to disable it due to following
errors:
> "Cannot find builtins library for the target architecture.
> Call Stack (most recent call first):
> /llvm_src/compiler-rt/lib/memprof/CMakeLists.txt:120  (add_compiler_rt_runtime)"